Adjusted language in m1_o3 resources objective
Added more details in the objectives tab, pointing the player to the documentation and academy webpages.
Added Discord link to the objectives tab

Added more details to the help section to be more clear where to get the answers from.

Added more helpful error messages in validation.
Added a validator for answering with too many framework SDKs.